requirements.txt
selenium==3.141.0
requests==2.24.0
PyYAML==5.3.1
